🚚rename : Clearly specify the language in the i18n file name
