+++ src/locales/en/en-page.json
@@ -1,0 +1,6 @@
+{
+  "startPage": "Start Page",
+  "404": {
+    "text": "Something went wrong. 😵"
+  }
+}
