--- src/locales/en/page.json
@@ -1,3 +1,0 @@
-{
-  "startPage": "Start Page"
-}
