+++ src/locales/ko/index.ts
@@ -1,4 +1,4 @@
-import page from "./page.json";
+import page from "./ko-page.json";
 
 export default {
   page,
