+++ src/locales/ko/ko-page.json
@@ -1,0 +1,6 @@
+{
+  "startPage": "Start Page",
+  "404": {
+    "text": "페이지가 존재하지 않습니다. 😵"
+  }
+}
