--- src/locales/ko/page.json
@@ -1,3 +1,0 @@
-{
-  "startPage": "시작 페이지"
-}
